pyproject.toml
[project]
name = "win-manager"
version = "0.1.0"
description = "A Windows system or window management tool"
readme = "README.md"
requires-python = ">=3.8.1"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["windows", "management", "system", "window"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pywin32>=306",
    "pyautogui>=0.9.54",
    "psutil>=5.9.0",
    "PyQt6>=6.6.0",
    "pynput>=1.7.6",
    "configparser>=5.3.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "colorama>=0.4.6",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
win-manager = "win_manager.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.1",
]
